Messaggio di conferma prima della cancellazione

+++ WebApplication/public/script.js
@@ -130,25 +130,27 @@ function loadFacilities_administrator(sport_center_id){
 //@param[id_sport_facility]: id of the sport facility that has to be deleted
 //@param[sport_center_id]: id of the sport center where the sport facility is
 function deleteSportFacility(id_sport_facility, sport_center_id){
-    var token = "empty";
-    var auth_level = "empty";
-    token = getCookie("token");
-    auth_level = getCookie("user_level");
+    if(confirm("Are you sure to delete the selected resource?")){
+        var token = "empty";
+        var auth_level = "empty";
+        token = getCookie("token");
+        auth_level = getCookie("user_level");
 
-    if(auth_level=="administrator"){
-        fetch('../api/v1/sport_facilities/'+id_sport_facility, {
-            method: 'DELETE',
-            headers: { "x-access-token": token },
-        })
-        .then((resp) => {
-            if(resp.status==403){
-                console.log("Authentication error");
-            }else{
-                loadFacilities_administrator(sport_center_id);
-            }
-        })
-    }else{
-        console.log("Authentication error");
+        if(auth_level=="administrator"){
+            fetch('../api/v1/sport_facilities/'+id_sport_facility, {
+                method: 'DELETE',
+                headers: { "x-access-token": token },
+            })
+            .then((resp) => {
+                if(resp.status==403){
+                    console.log("Authentication error");
+                }else{
+                    loadFacilities_administrator(sport_center_id);
+                }
+            })
+        }else{
+            console.log("Authentication error");
+        }
     }
 }
 //...
